requirements.txt
requests~=2.21.0
Flask==1.1.1
beautifulsoup4==4.8.2
langdetect==1.0.7
python-dotenv==0.10.5